Eagerly fetch the ignore status for the open editor.
