+++ src/main/kotlin/com/sourcegraph/cody/ignore/CommandPanelIgnoreBanner.kt
@@ -14,58 +14,29 @@ import com.sourcegraph.Icons
 import com.sourcegraph.common.CodyBundle
 import java.awt.Dimension
 
-class CommandPanelIgnoreBanner(val project: Project) : NonOpaquePanel(), IgnoreOracle.FocusedFileIgnorePolicyListener {
+class CommandPanelIgnoreBanner() : NonOpaquePanel() {
   init {
     ApplicationManager.getApplication().assertIsDispatchThread()
-  }
 
-  private val banner = EditorNotificationPanel().apply {
-    text = CodyBundle.getString("ignore.sidebar-panel-ignored-file.text")
-    createActionLabel(
-      CodyBundle.getString("ignore.sidebar-panel-ignored-file.learn-more-cta"),
-      { BrowserUtil.browse(CODY_IGNORE_DOCS_URL) }, false)
-    icon(Icons.CodyLogoSlash)
-  }
+    add(EditorNotificationPanel().apply {
+      text = CodyBundle.getString("ignore.sidebar-panel-ignored-file.text")
+      createActionLabel(
+        CodyBundle.getString("ignore.sidebar-panel-ignored-file.learn-more-cta"),
+        { BrowserUtil.browse(CODY_IGNORE_DOCS_URL) }, false
+      )
+      icon(Icons.CodyLogoSlash)
+    })
 
-  init {
-    addHierarchyListener {
-      if (!project.isDisposed) {
-        if (it.component.isShowing) {
-          IgnoreOracle.getInstance(project).addListener(this)
-        } else {
-          IgnoreOracle.getInstance(project).removeListener(this)
-        }
-      }
-    }
-  }
-
-  private fun update(policy: IgnorePolicy) {
-     when (policy) {
-      IgnorePolicy.USE -> {
-        remove(banner)
-        border = JBEmptyBorder(0)
-      }
-      IgnorePolicy.IGNORE -> {
-        add(banner)
-
-        // These colors cribbed from EditorComposite, createTopBottomSideBorder
-        val scheme = EditorColorsManager.getInstance().globalScheme
-        val borderColor =
-          scheme.getColor(EditorColors.SEPARATOR_ABOVE_COLOR) ?: scheme.getColor(EditorColors.TEARLINE_COLOR)
-        border = SideBorder(borderColor, SideBorder.TOP or SideBorder.BOTTOM)
-      }
-    }
+    // These colors cribbed from EditorComposite, createTopBottomSideBorder
+    val scheme = EditorColorsManager.getInstance().globalScheme
+    val borderColor =
+      scheme.getColor(EditorColors.SEPARATOR_ABOVE_COLOR) ?: scheme.getColor(EditorColors.TEARLINE_COLOR)
+    border = SideBorder(borderColor, SideBorder.TOP or SideBorder.BOTTOM)
   }
 
   override fun getMaximumSize(): Dimension {
     val size = super.getMaximumSize()
     size.height = preferredSize.height
     return size
-  }
-
-  override fun focusedFileIgnorePolicyChanged(policy: IgnorePolicy) {
-    runInEdt {
-      update(policy)
-    }
   }
 }
